XL: Relax write quorum further to N/2 + 1.
This changes behavior in some parts of the code
as well address it.

Fixes #2016
